Updated Main
Due to how Flatpak is a security jail for processes, added detection if we are in a flatpak, if we are in flatpak, we use new custom launch_backend function called flatpak_launch_backend.

flatpak_launch_backend(): Does the same thing as launch_backend(), except instead of normal invocation of the backend.py script, we launch the backend.py script through flatpak-spawn --host, to have the backend run in the system context, which should solve issues with access to /tmp folder, as well as process information for watching log files.