Update Addon
Added Timeout on displays when we haven't received anything for 5 seconds, clear the display.
Fixed displaying of data to be more properly aligned with what is displayed in the HUD on the buttons.
Added Action Toggle Hud which will toggle the Mango Hud.
Added Control Script to communicate with MangoHud through socket.
Changed Log Reader to pass self along, so that we can look it up.
Added class struct to handle multiple data points that are needed to manage.
Added Automatically hidding the Hud after we get our first line of data.