Fixed some crashing of backend
Added true/false return on control() to show if it succeeds or fails.  Attempts to execute command once every 100 ms, for up to 5 seconds, before failing.
Changed toggle_hud() to return true or false if we successfully issue the command through the socket.  (Automatically returns true if there's no socket to send to in the first place)
Added toggling to show the error when trying to toggle a hud, and it fails.